requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autocorrect==2.6.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
editdistance==0.8.1
et_xmlfile==2.0.0
Flask==3.1.0
flatbuffers==2.0.7
fonttools==4.57.0
frozenlist==1.6.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
libretranslatepy==2.1.1
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
multidict==6.4.3
nltk==3.9.1
numpy==1.23.5
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
telepot==12.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
translate==3.6.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
